+++ NEWS.md
@@ -8,6 +8,7 @@
 * Avoid newly deprecated arguments and functions in `crew` 0.5.0.9003 (https://github.com/wlnadau/crew/issues/131).
 * Allow `tar_read()` etc. inside a pipeline whenever it uses a different data store (#1158, @MilesMcBain).
 * Set `seed = FALSE` in `future::future()` (#1166, @svraka).
+* Add a new `physics` argument to `tar_visnetwork()` and `tar_glimpse()` (#925, @Bdblodgett-usgs).
 
 # targets 1.3.0